Removed Unnecessary Codes

+++ simElectricity/src/main/java/simElectricity/EnergyNet/Matrix/Gaussian.java
@@ -20,7 +20,7 @@
 package simElectricity.EnergyNet.Matrix;
 
 public class Gaussian implements IMatrixResolver{
-    public static final double EPSILON = (double) 1e-10;
+    public static final double EPSILON = 1e-10;
 
     double[][] matrix;
     int currentRow;
